+++ tests/test_hmf_integrals.py
@@ -1,0 +1,25 @@
+import py21cmfast as p21c
+
+# we should test the HMF values and integrals
+#   at a range of densities vs known quantities
+#   maybe import hmf to verify
+
+
+def test_cmf_integrals():
+    # take a bunch of (sigma,delta) and compare against integrals of HMFs
+    pass
+
+
+def test_cmf_values():
+    # take a bunch of (sigma,delta,M) and compare against values of HMFs
+    pass
+
+
+def test_umf_integrals():
+    # take a bunch of (z) and compare against integrals of HMFs
+    pass
+
+
+def test_umf_values():
+    # take a bunch of (z,M) and compare against values of HMFs
+    pass
